update-vm-scanning.adoc
Clarity to ideally avoid customer confusion in the future. It's not explicity stated that this functionality is how it works or is restricted, but testing has shown it is the case. Can provide error logs/examples upon request.
Scoping for this is 1:1 . 1 VPC, 1 Subnet, PER rule/scope for VM image scanning .