Fix PCSUP-7410 (#152)
There's currently an issue with the HTML transform, which doesn't properly close the div for the note when a note contains a codeblock element.

Refactor the note so that it doesn't use a codeblock, and so that the page renders properly.

(cherry picked from commit 94fcb7931b1b3026f517dbc1d8f2874baa8c8ec7)