clarify user_roles.adoc
this shows up in the saas docs, but none of these roles are relevant to SaaS in any way. None of them are available. We need to clarify that piece and/or remove this page entirely as it has no value technically for SaaS deployments. Perhaps PM team or platform team can comment or something