Add section on GKE Autopilot docs
Add section on GKE Autopilot docs that explicity states "GKE AUTOPILOT DEFENDERS WILL ONLY WORK WITH OFFICIAL TWISTLOCK REGISTRY, NOT CUSTOMER PROVIDED REGISTRY" in regards to the defender image. This is causing a lot of issues with my customers and should be stated in the documentation. I didn't know this until I dug through a bunch of PCSUPs.