[CSPM AG] Update to match AWS UI change (#1046)
Comment on Slack on March 27, 2022 from Dennis Schmidt. Changes 
On Step 4.8, the AWS console changed to default to the "advanced event selectors".  To follow the instructions as is, the user would first need to click the button "Switch to basic event selectors" to see the screen in the doc.  If the user tries to do the same steps with the advanced selectors, it will not work as expected.  It's clear from the screenshot that the basic selector is required - perhaps just note to indicate to switch to basic if on advanced.