Update prisma-cloud-licenses.adoc
revised the details on 

Credit consumption is measured hourly and the hourly samples are averaged to create daily samples. The credit usage for a specified time range uses the appropriate hourly, daily or monthly average. If there is less than 30 days of data available, the average is calculated using the days available

This inline with https://github.com/PaloAltoNetworks/prisma-cloud-docs/pull/1210